requirements.txt
Django~=3.0
PyJWT~=1.7.1
cryptography~=2.8
pip~=20.1.1
wheel~=0.34.2
keyring~=18.0.1
pymongo~=3.10.1
pytz~=2019.3
sqlparse~=0.2.4
mysqlclient~=1.4.6
asyncio~=3.4.3
asgiref~=3.2.10
PyYAML~=5.3.1
Jinja2~=2.11.2
djongo~=1.3.3
MarkupSafe~=1.1.0
certifi~=2020.4.5.1
chardet~=3.0.4
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
requests~=2.23.0
click~=7.1.2
uritemplate~=3.0.1
monotonic~=1.5
idna~=2.9
urllib3~=1.25.9
simplejson~=3.16.0
setuptools~=45.2.0
djangorestframework~=3.10.0
Pygments~=2.3.1